Refactor linear-cost splitting into struct
